Use a prebuilt Stardoc release binary
Instead of forcing users to depend on @io_bazel so that Stardoc
is built from source for each invocation, a prebuilt Stardoc binary will be included
in this repository, and updated on release cuts.

This also eliminates buggy behavior when Stardoc and @io_bazel repositories
are out of sync in a user's workspace.

Closes #170